Did work on static Noise wrapper and added CoherentNoise (rather haphazardly)
CoherentNoise is an old 'library' that offers support for world generation. Seeing i have no idea what the hell i am doing half the time, I just dropped the CoherentNoise solution into the external folder in its own subfolder. Unity sees it and so do my scripts, so ill just try to use it as it.